editor/backend/lead/requirements.txt
Django==2.2.18
django_filter==2.4.0
djangorestframework==3.12.2
django-cors-headers==3.5.0
psycopg2==2.8.6
